+++ slh_dsa.v
@@ -202,7 +202,7 @@ fn slh_sign(c Context, m []u8, cx []u8, sk Sk, opt SignerOpts) ![]u8 {
 	// 𝑀′ ← toByte(0, 1) ∥ toByte(|𝑐𝑡𝑥|, 1) ∥ 𝑐𝑡𝑥 ∥ m
 	mut msg := []u8{}
 	msg << to_byte(0, 1)
-	msg << to_byte(cx.len, 1)
+	msg << to_byte(u64(cx.len), 1)
 	msg << cx
 	msg << m
 
@@ -291,7 +291,7 @@ fn slh_verify(c Context, m []u8, sig []u8, cx []u8, pk Pk) !bool {
 	// 𝑀′ ← toByte(0, 1) ∥ toByte(|𝑐𝑡𝑥|, 1) ∥ 𝑐𝑡𝑥 ∥ m
 	mut msg := []u8{}
 	msg << to_byte(0, 1)
-	msg << to_byte(cx.len, 1)
+	msg << to_byte(u64(cx.len), 1)
 	msg << cx
 	msg << m
