+++ wotsp.v
@@ -32,7 +32,7 @@ fn wots_pkgen(c Context, sk_seed []u8, pk_seed []u8, mut addr Address) ![]u8 {
 	// skADRS.setTypeAndClear(WOTS_PRF)
 	sk_addr.set_type_and_clear(.wots_prf)
 	// skADRS.setKeyPairAddress(ADRS.getKeyPairAddress())
-	sk_addr.set_keypair_address(addr.get_keypair_address())
+	sk_addr.set_keypair_address(u32(addr.get_keypair_address()))
 
 	// gets wotsp length from the current context
 	wots_len := c.prm.wots_len()
@@ -40,11 +40,11 @@ fn wots_pkgen(c Context, sk_seed []u8, pk_seed []u8, mut addr Address) ![]u8 {
 	mut tmp := []u8{}
 	for i := 0; i < wots_len - 1; i++ {
 		// skADRS.setChainAddress(𝑖)
-		sk_addr.set_chain_address(i)
+		sk_addr.set_chain_address(u32(i))
 		// compute secret value for chain i, 𝑠𝑘 ← PRF(PK.seed, SK.seed, skADRS)
 		sk := c.prf(pk_seed, sk_seed, sk_addr)!
 		// ADRS.setChainAddress(𝑖)
-		addr.set_chain_address(i)
+		addr.set_chain_address(u32(i))
 		// compute public value for chain 𝑖, 𝑡𝑚𝑝[𝑖] ← chain(𝑠𝑘, 0, 𝑤 − 1, PK.seed, ADRS)
 		tmp_i := chain(c, sk, 0, w - 1, pk_seed, mut addr)!
 		tmp << tmp_i
@@ -55,7 +55,7 @@ fn wots_pkgen(c Context, sk_seed []u8, pk_seed []u8, mut addr Address) ![]u8 {
 	wots_pk_addr.set_type_and_clear(.wots_pk)
 	// wotspkADRS.setKeyPairAddress(ADRS.getKeyPairAddress())
 	// TODO: remove int cast
-	wots_pk_addr.set_keypair_address(int(addr.get_keypair_address()))
+	wots_pk_addr.set_keypair_address(u32(addr.get_keypair_address()))
 	// compress public key, 𝑝𝑘 ← T𝑙𝑒𝑛(PK.seed, wotspkADRS,𝑡𝑚𝑝)
 	pk := c.tlen(pk_seed, wots_pk_addr, tmp)!
 
@@ -70,7 +70,7 @@ fn wots_pkgen(c Context, sk_seed []u8, pk_seed []u8, mut addr Address) ![]u8 {
 // Input: Message 𝑀, secret seed SK.seed, public seed PK.seed, address ADRS.
 // Output: WOTS+ signature 𝑠𝑖𝑔.
 fn wots_sign(c Context, m []u8, sk_seed []u8, pk_seed []u8, mut addr Address) ![]u8 {
-	mut csum := 0
+	mut csum := u64(0)
 	// convert message to base w, ie, 𝑚𝑠𝑔 ← base_2b(𝑀, 𝑙𝑔𝑤, 𝑙𝑒𝑛1)
 	len1 := c.prm.len1()
 	mut msgs := base_2exp_b(m, c.prm.lgw, len1)
@@ -78,11 +78,11 @@ fn wots_sign(c Context, m []u8, sk_seed []u8, pk_seed []u8, mut addr Address) ![
 	// compute checksum
 	for i := 0; i < len1 - 1; i++ {
 		// 𝑐𝑠𝑢𝑚 ← 𝑐𝑠𝑢𝑚 + 𝑤 − 1 − 𝑚𝑠𝑔[𝑖]
-		csum += int(w - 1 - msgs[i])
+		csum += w - 1 - msgs[i]
 	}
 	// for 𝑙𝑔𝑤 = 4, left shift by 4, its only values supported in this module
 	// 𝑐𝑠𝑢𝑚 ← 𝑐𝑠𝑢𝑚 ≪ ((8 − ((𝑙𝑒𝑛2 ⋅ 𝑙𝑔𝑤) mod 8)) mod 8)
-	csum <<= (8 - ((len2 * c.prm.lgw) % 8)) % 8
+	csum <<= u64((8 - ((len2 * c.prm.lgw) % 8)) % 8)
 
 	// convert to base w, 𝑚𝑠𝑔 ← 𝑚𝑠𝑔 ∥ base_2b (toByte (𝑐𝑠𝑢𝑚, ⌈(𝑙𝑒𝑛2*𝑙𝑔𝑤)/8⌉) , 𝑙𝑔𝑤, 𝑙𝑒𝑛2)
 	mlen := cdiv(len2 * c.prm.lgw, 8)
@@ -95,15 +95,15 @@ fn wots_sign(c Context, m []u8, sk_seed []u8, pk_seed []u8, mut addr Address) ![
 	sk_addr.set_type_and_clear(.wots_prf)
 	// skADRS.setKeyPairAddress(ADRS.getKeyPairAddress())
 	// TODO: handle int > larger than max_int
-	sk_addr.set_keypair_address(int(addr.get_keypair_address()))
+	sk_addr.set_keypair_address(u32(addr.get_keypair_address()))
 
 	mut sig := []u8{}
 	for i := 0; i < c.prm.wots_len(); i++ {
 		// skADRS.setChainAddress(𝑖)
-		sk_addr.set_chain_address(i)
+		sk_addr.set_chain_address(u32(i))
 		// compute chain 𝑖 secret value, 𝑠𝑘 ← PRF(PK.seed, SK.seed, skADRS)
 		sk := c.prf(pk_seed, sk_seed, sk_addr)!
-		addr.set_chain_address(i)
+		addr.set_chain_address(u32(i))
 		// compute chain 𝑖 signature value, 𝑠𝑖𝑔[𝑖] ← chain(𝑠𝑘, 0, 𝑚𝑠𝑔[𝑖], PK.seed, ADRS)
 		sig_i := chain(c, sk, 0, int(msgs[i]), pk_seed, mut addr)!
 		sig << sig_i
@@ -118,7 +118,7 @@ fn wots_sign(c Context, m []u8, sk_seed []u8, pk_seed []u8, mut addr Address) ![
 // Input: WOTS+ signature 𝑠𝑖𝑔, message 𝑀, public seed
 // Output: WOTS+ public key 𝑝𝑘𝑠𝑖𝑔 derived from 𝑠𝑖𝑔.
 fn wots_pkfromsig(c Context, sig []u8, m []u8, pk_seed []u8, mut addr Address) ![]u8 {
-	mut csum := 0
+	mut csum := u64(0)
 	// convert message to base w, ie, 𝑚𝑠𝑔 ← base_2b(𝑀, 𝑙𝑔𝑤, 𝑙𝑒𝑛1)
 	len1 := c.prm.len1()
 	mut msgs := base_2exp_b(m, c.prm.lgw, len1)
@@ -126,11 +126,11 @@ fn wots_pkfromsig(c Context, sig []u8, m []u8, pk_seed []u8, mut addr Address) !
 	// compute checksum
 	for i := 0; i < len1 - 1; i++ {
 		// 𝑐𝑠𝑢𝑚 ← 𝑐𝑠𝑢𝑚 + 𝑤 − 1 − 𝑚𝑠𝑔[𝑖]
-		csum += int(w - 1 - msgs[i])
+		csum += w - 1 - msgs[i]
 	}
 	// for 𝑙𝑔𝑤 = 4, left shift by 4, its only values supported in this module
 	// 𝑐𝑠𝑢𝑚 ← 𝑐𝑠𝑢𝑚 ≪ ((8 − ((𝑙𝑒𝑛2 ⋅ 𝑙𝑔𝑤) mod 8)) mod 8)
-	csum <<= (8 - ((len2 * c.prm.lgw) % 8)) % 8
+	csum <<= u64((8 - ((len2 * c.prm.lgw) % 8)) % 8)
 
 	// convert to base w, 𝑚𝑠𝑔 ← 𝑚𝑠𝑔 ∥ base_2b (toByte (𝑐𝑠𝑢𝑚, ⌈(𝑙𝑒𝑛2*𝑙𝑔𝑤)/8⌉) , 𝑙𝑔𝑤, 𝑙𝑒𝑛2)
 	mlen := cdiv(len2 * c.prm.lgw, 8)
@@ -140,7 +140,7 @@ fn wots_pkfromsig(c Context, sig []u8, m []u8, pk_seed []u8, mut addr Address) !
 	mut tmp := []u8{}
 	for i := 0; i < c.prm.wots_len(); i++ {
 		// ADRS.setChainAddress(𝑖)
-		addr.set_chain_address(i)
+		addr.set_chain_address(u32(i))
 		// 𝑡𝑚𝑝[𝑖] ← chain(𝑠𝑖𝑔[𝑖], 𝑚𝑠𝑔[𝑖], 𝑤 − 1 − 𝑚𝑠𝑔[𝑖], PK.seed, ADRS)
 		x := sig[i * c.prm.n..(i + 1) * c.prm.n]
 		next_chain := chain(c, x, int(msgs[i]), int(w - 1 - msgs[i]), pk_seed, mut addr)!
@@ -154,7 +154,7 @@ fn wots_pkfromsig(c Context, sig []u8, m []u8, pk_seed []u8, mut addr Address) !
 	wots_pk_addr.set_type_and_clear(.wots_pk)
 	// wotspkADRS.setKeyPairAddress(ADRS.getKeyPairAddress())
 	// TODO: remove int casts ??
-	wots_pk_addr.set_keypair_address(int(addr.get_keypair_address()))
+	wots_pk_addr.set_keypair_address(u32(addr.get_keypair_address()))
 	// 𝑝𝑘𝑠𝑖𝑔 ← T𝑙𝑒𝑛(PK.seed, wotspkADRS,𝑡𝑚𝑝)
 	pk_sig := c.tlen(pk_seed, wots_pk_addr, tmp)!
